feat(build): use gulp with typescript
