+++ package.json
@@ -101,6 +101,7 @@
     "@angular/cli": "1.7.1",
     "@angular/compiler-cli": "^5.2.6",
     "@angular/language-service": "^5.2.6",
+    "@types/gulp": "4.0.5",
     "@types/jasmine": "2.8.3",
     "@types/jasminewd2": "2.0.3",
     "@types/leaflet": "1.0.60",
